more 10.5 support
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2826 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ CalciumScoringWindowController.m
@@ -571,9 +571,9 @@ enum ctTypes {ElectronCTType, MultiSliceCTType};
 - (NSString *)patientsAge{
 	NSCalendarDate *dob = [[[_viewer currentStudy] valueForKey:@"dateOfBirth"] dateWithCalendarFormat:nil timeZone:nil];
 	NSCalendarDate *studyDate = [[[_viewer currentStudy] valueForKey:@"date"] dateWithCalendarFormat:nil timeZone:nil];
-	int years;
-	int days;
-	int months;
+	NSInteger years;
+	NSInteger days;
+	NSInteger months;
 	if (dob && studyDate) {
 		[studyDate years:&years
 		 months:&months
